+++ scripts/install.sh
@@ -29,8 +29,8 @@ if [ -z "$KIME_ICON_DIR" ]; then
     KIME_ICON_DIR=usr/share/icons
 fi
 
-if [ -z "$KIME_APP_DIR" ]; then
-    KIME_APP_DIR=usr/share/applications
+if [ -z "$KIME_AUTOSTART_DIR" ]; then
+    KIME_AUTOSTART_DIR=etc/xdg/autostart
 fi
 
 if [ -z "$KIME_LIB_DIR" ]; then
@@ -80,7 +80,7 @@ if [ $KIME_INSTALL_HEADER -eq "1" ]; then
 fi
 
 install -Dm644 $KIME_OUT/default_config.yaml -T "$PREFIX/$KIME_CONFIG_DIR/config.yaml"
-install -Dm644 $KIME_OUT/*.desktop -t "$PREFIX/$KIME_APP_DIR"
+install -Dm644 $KIME_OUT/*.desktop -t "$PREFIX/$KIME_AUTOSTART_DIR"
 install -Dm644 $KIME_OUT/icons/64x64/*.png -t "$PREFIX/$KIME_ICON_DIR/hicolor/64x64/apps"
 install -Dm755 $KIME_OUT/libkime_engine.so -t "$PREFIX/$KIME_LIB_DIR"